Fixed delete button in reservations screen being on a new line
